Included ActiveRecord for ORM, this way it won't matter if we switch from PG to MySQL or vice versa. It is also easier to communicate with database and to query.
Included Slim for a small and simple RESTful server with logging support.

ActiveRecord has been patched as it had no way to output/serialize
entities as JSON.

Created a server application to handle /ratings, /pluginStatus GET resources
and /ratings POST.

There is also a sample application that shows how to use this application
for keeping status of plug-in compatibility. This sample uses only Twitter
bootstrap and Backbone.js/JQuery.

The database model was created based on the existing code of infra-rating,
and it is not normalized. There is a simple index, but it is not clear
whether that will be enough or not.

Before running the application, check the database connection, Slim mode
(development, production, etc) and if the log leves are fine.